Add coverage reporting
In travis, add distinct build stages for testing with coverage report
and deploying with a build type that does not generate coverage
information in downstream projects.

Uses codecov.io for coverage reporting, which needs to be enabled on the
repo and the token must be specified in the travis environment.